Refactor job_dispatcher
