+++ taipy/core/_scheduler/_job_dispatcher.py
@@ -8,48 +8,43 @@
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
-
-from concurrent.futures import ProcessPoolExecutor
-from functools import partial
+from abc import abstractmethod
 from typing import Any, List
 
+from taipy.core.common._taipy_logger import _TaipyLogger
 from taipy.core.common.alias import JobId
 from taipy.core.config import Config
 from taipy.core.data._data_manager_factory import _DataManagerFactory
 from taipy.core.data.data_node import DataNode
 from taipy.core.exceptions.exceptions import DataNodeWritingError
-from taipy.core.job._job_manager_factory import _JobManagerFactory
 from taipy.core.job.job import Job
 from taipy.core.task.task import Task
 
 
 class _JobDispatcher:
-    """Manages executors and dispatch jobs (instances of `Job^` class) on them."""
+    """Manages job dispatching (instances of `Job^` class) on executors."""
 
     def __init__(self):
-        self._executor = ProcessPoolExecutor(Config.job_config.nb_of_workers or 1)
-        self._nb_available_workers = self._executor._max_workers  # type: ignore
+        self.__logger = _TaipyLogger._get_logger()
 
+    @abstractmethod
     def _can_execute(self) -> bool:
-        """Returns True if a worker is available for a new run."""
-        return self._nb_available_workers > 0
+        """Returns True if the dispatcher have resources to execute a new job."""
+        raise NotImplementedError
 
+    @abstractmethod
     def _dispatch(self, job: Job):
-        """Dispatches the given `Job^` on an available worker for execution.
+        """
+        Dispatches the given `Job^` on an available worker for execution.
 
         Parameters:
             job (Job^): The job to submit on an executor with an available worker.
         """
-        self._nb_available_workers -= 1
-        future = self._executor.submit(
-            self._run_wrapped_function, Config.global_config.storage_folder, job.id, job.task
-        )
-        future.add_done_callback(self.__release_worker)
-        future.add_done_callback(partial(self.__update_status, job))
+        raise NotImplementedError
 
     @classmethod
     def _run_wrapped_function(cls, storage_folder: str, job_id: JobId, task: Task):
-        """)
+        """
         Reads inputs, execute function, and write outputs.
 
         Parameters:
@@ -67,14 +62,6 @@ class _JobDispatcher:
             return cls.__write_data(outputs, results, job_id)
         except Exception as e:
             return [e]
-
-    def __release_worker(self, _):
-        self._nb_available_workers += 1
-
-    @staticmethod
-    def __update_status(job: Job, ft):
-        job.update_status(ft)
-        _JobManagerFactory._build_manager()._set(job)
 
     @classmethod
     def __read_inputs(cls, inputs: List[DataNode]) -> List[Any]:
@@ -106,8 +93,3 @@ class _JobDispatcher:
             raise DataNodeWritingError("Error: wrong number of result or task output")
 
         return _results
-
-    @staticmethod
-    def __unlock_edit_on_outputs(job):
-        for dn in job.task.output.values():
-            dn.unlock_edit(job_id=job.id)
